requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==13.1
openai>=1.58.1
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.28.1
python-multipart==0.0.17
